requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-colorful==1.3
django-crispy-forms==1.9.2
django-multiselectfield==0.1.12
djangorestframework==3.11.1
idna==2.10
MouseInfo==0.1.3
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
attrs==20.2.0